change project crates struct
