+++ Cargo.toml
@@ -1,37 +1,37 @@
-[package]
-name = "justpic"
-version = "0.1.0"
-edition = "2024"
+[workspace]
+resolver = "2"
+members = [
+	"crates/server",
+	"crates/core/*"
+]
 
-[dependencies]
+[workspace.dependencies]
 actix-web = "4.11.0"
-
-tokio = {version = "1.47.1", features = ["sync", "rt-multi-thread", "macros"]}
-
-sqlx = {version = "0.8.5", features = ["runtime-tokio", "postgres", "macros", "time", "uuid"]}
-
-deadpool-redis = {version = "0.22.0"}
-redis = {version = "0.32.4", features = ["tokio-comp"]}
-
-dotenvy = { version = "0.15.7"}
-
-serde = {version = "1.0.219", features = ["derive"]}
-serde_json = {version = "1.0.140"}
-
-time = { version = "0.3.41", features = ["serde", "serde-human-readable"] }
-
+tokio = "1.47.1"
+sqlx = "0.8.5"
+deadpool-redis = "0.22.0"
+redis = "0.32.4"
+dotenvy = "0.15.7"
+serde = "1.0.219"
+serde_json = "1.0.140"
+time = "0.3.41"
 tracing = "0.1.40"
-tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
+tracing-subscriber = "0.3.18"
+argon2 = "0.5.3"
+rand = "0.9.2"
+uuid = "1.17.0"
+base64 = "0.22.1"
+derive_more = "2.0.1"
+validator = "0.19"
 
-argon2 = { version = "0.5.3", features = ["std"] }
-
-rand = {version = "0.9.2"}
-uuid = {version = "1.17.0", features = ["v4", "serde"]}
-
-# change to HEX
-base64 = {version = "0.22.1"} 
-
-# Custom justpic-team utilities
 boar-utils = {git = "https://github.com/justpic/boar-utils", branch = "master"}
 
-derive_more = {version = "2.0.1", features = ["from", "display"]}
+[profile.release]
+opt-level = "s"
+strip = true
+lto = true
+panic = "abort"
+codegen-units = 1
+
+[profile.dev.package.sqlx-macros]
+opt-level = 3
