+++ crates/core/database/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "justpic-database"
+edition = "2024"
+version = "0.1.0"
+
+[dependencies]
+justpic-result = {path = "../result"}
+
+tokio = {workspace = true, features = ["sync", "rt-multi-thread", "macros"]}
+
+sqlx = {workspace = true, features = ["runtime-tokio", "postgres", "macros", "time", "uuid"]}
+
+deadpool-redis.workspace = true
+redis = {workspace = true, features = ["tokio-comp"]}
+
+dotenvy.workspace = true
+
+serde = {workspace = true, features = ["derive"]}
+serde_json.workspace = true
+
+time = { workspace = true, features = ["serde", "serde-human-readable"] }
+
+tracing.workspace = true
+tracing-subscriber = { workspace = true, features = ["env-filter"] }
+
+uuid = {workspace = true, features = ["v4", "serde"]}
