+++ crates/server/src/routes/v1/auth/mod.rs
@@ -1,0 +1,5 @@
+use actix_web::web;
+
+pub fn config(cfg: &mut web::ServiceConfig) {
+    cfg;
+}
