+++ crates/server/src/routes/v1/mod.rs
@@ -1,0 +1,9 @@
+use actix_web::web;
+
+pub mod auth;
+pub mod users;
+
+pub fn config(cfg: &mut web::ServiceConfig) {
+    cfg.service(web::scope("/auth").configure(auth::config))
+        .service(web::scope("/users").configure(users::config));
+}
